leomega parallel

+++ src/misc/analysis.jl
@@ -74,12 +74,13 @@ function rms_ellipsoid(v1,a,b,c,ngrid)
     uz =  real.([v[3](Mire.x=>xt,Mire.y=>yt,Mire.z=>zt) for (xt,yt,zt) in zip(X,Y,Z)])
     √mean(ux.^2+uy.^2+uz.^2)
 end
+
 function rmsle(N,vs,u,a,b,c,ngrid=40)
     v_ile=Mire.eigenvel(N,vs,u,a,b,c,norm=false)
     v_geo,v_ageo = geoageo(v_ile,a,b,c)
     rms2=rms_eqplane(v_geo,a,b,ngrid)
     rms3=rms_eqplane(v_ageo,a,b,ngrid)
-    return rms3/rms2
+    return Float64(rms3/rms2)
 end
 
 function meshgrid(vx::AbstractVector{T}, vy::AbstractVector{T},
